+++ app/src/api/controllers/buildingController.ts
@@ -1,7 +1,8 @@
 import express from 'express';
+
+import asyncController from '../routes/asyncController';
 import * as buildingService from '../services/building';
 import * as userService from '../services/user';
-import asyncController from '../routes/asyncController';
 
 
 // GET buildings
