+++ app/src/api/routes/asyncController.ts
@@ -1,4 +1,4 @@
-import { Request, Response, NextFunction } from 'express';
+import { NextFunction, Request, Response } from 'express';
 
 /**
  * A wrapper for controller functions that return a Promise, enabling them to be used with Express
