+++ app/src/frontend/building/data-components/checkbox-data-entry.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 
-import { DataTitleCopyable } from './data-title';
 import { BaseDataEntryProps } from './data-entry';
+import { DataTitleCopyable } from './data-title';
 
 interface CheckboxDataEntryProps extends BaseDataEntryProps {
     value: boolean;
