+++ app/src/frontend/building/data-components/numeric-data-entry.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 
-import { DataTitleCopyable } from './data-title';
 import { BaseDataEntryProps } from './data-entry';
+import { DataTitleCopyable } from './data-title';
 
 
 interface NumericDataEntryProps extends BaseDataEntryProps {
