+++ app/src/frontend/building/data-components/textbox-data-entry.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 
-import { DataTitleCopyable } from './data-title';
 import { BaseDataEntryProps } from './data-entry';
+import { DataTitleCopyable } from './data-title';
 
 interface TextboxDataEntryProps extends BaseDataEntryProps {
     value: string;
