+++ app/src/frontend/building/data-containers/sustainability.tsx
@@ -1,10 +1,10 @@
 import React, { Fragment } from 'react';
 
-import withCopyEdit from '../data-container';
-import DataEntry from '../data-components/data-entry';
-import SelectDataEntry from '../data-components/select-data-entry';
-import NumericDataEntry from '../data-components/numeric-data-entry';
 import { dataFields } from '../../data_fields';
+import NumericDataEntry from '../data-components/numeric-data-entry';
+import SelectDataEntry from '../data-components/select-data-entry';
+import withCopyEdit from '../data-container';
+
 import { CategoryViewProps } from './category-view-props';
 
 const EnergyCategoryOptions = ["A", "B", "C", "D", "E", "F", "G"];
