+++ app/src/frontend/building/data-containers/team.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment } from 'react';
 
 import withCopyEdit from '../data-container';
-import DataEntry from '../data-components/data-entry';
+
 import { CategoryViewProps } from './category-view-props';
 
 /**
