+++ app/src/frontend/building/edit-history/building-edit-summary.tsx
@@ -1,7 +1,9 @@
 import React from 'react';
-import { EditHistoryEntry } from '../models/edit-history-entry';
-import { arrayToDictionary, parseDate } from '../../helpers';
+
 import { dataFields } from '../../data_fields';
+import { arrayToDictionary, parseDate } from '../../helpers';
+import { EditHistoryEntry } from '../../models/edit-history-entry';
+
 import { CategoryEditSummary } from './category-edit-summary';
 
 import './building-edit-summary.css';
