+++ app/src/frontend/components/logo.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+
 import './logo.css';
 
 interface LogoProps {
