+++ app/src/frontend/map/map.tsx
@@ -1,15 +1,16 @@
-import React, { Component, Fragment } from 'react';
-import { Map, TileLayer, ZoomControl, AttributionControl, GeoJSON } from 'react-leaflet-universal';
 import { GeoJsonObject } from 'geojson';
-
-import '../../../node_modules/leaflet/dist/leaflet.css'
-import './map.css'
+import React, { Component, Fragment } from 'react';
+import { AttributionControl, GeoJSON, Map, TileLayer, ZoomControl } from 'react-leaflet-universal';
 
 import { HelpIcon } from '../components/icons';
+import { Building } from '../models/building';
+
 import Legend from './legend';
 import SearchBox from './search-box';
 import ThemeSwitcher from './theme-switcher';
-import { Building } from '../models/building';
+
+import '../../../node_modules/leaflet/dist/leaflet.css'
+import './map.css'
 
 const OS_API_KEY = 'NVUxtY5r8eA6eIfwrPTAGKrAAsoeI9E9';
