+++ app/src/frontend/pages/contributor-agreement.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import InfoBox from '../components/info-box';
-
 const ContributorAgreementPage : React.SFC<any> = () => (
     <article>
         <section className='main-col'>
