+++ app/src/frontend/pages/privacy-policy.tsx
@@ -1,8 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 
-import InfoBox from '../components/info-box';
-
 const PrivacyPolicyPage: React.SFC<any> = () => (
     <article>
         <section className='main-col'>
